+++ tubular/nominal.py
@@ -55,6 +55,9 @@ class BaseNominalTransformer(BaseTransformer):
     FITS: bool
         class attribute, indicates whether transform requires fit to be run first
 
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     Example:
     --------
     >>> BaseNominalTransformer(
@@ -67,6 +70,8 @@ class BaseNominalTransformer(BaseTransformer):
     polars_compatible = True
 
     jsonable = False
+
+    lazyframe_compatible = False
 
     FITS = False
 
@@ -279,6 +284,9 @@ class GroupRareLevelsTransformer(BaseTransformer, WeightColumnMixin):
     FITS: bool
         class attribute, indicates whether transform requires fit to be run first
 
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     Example:
     --------
     >>> GroupRareLevelsTransformer(
@@ -291,6 +299,8 @@ class GroupRareLevelsTransformer(BaseTransformer, WeightColumnMixin):
     """
 
     polars_compatible = True
+
+    lazyframe_compatible = False
 
     jsonable = True
 
@@ -798,6 +808,9 @@ class MeanResponseTransformer(
     FITS: bool
         class attribute, indicates whether transform requires fit to be run first
 
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     Example:
     --------
     >>> MeanResponseTransformer(
@@ -809,6 +822,8 @@ class MeanResponseTransformer(
     """
 
     polars_compatible = True
+
+    lazyframe_compatible = False
 
     jsonable = False
 
@@ -1479,6 +1494,9 @@ class OneHotEncodingTransformer(
     FITS: bool
         class attribute, indicates whether transform requires fit to be run first
 
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     Example:
     --------
     >>> OneHotEncodingTransformer(
@@ -1488,6 +1506,8 @@ class OneHotEncodingTransformer(
     """
 
     polars_compatible = True
+
+    lazyframe_compatible = False
 
     jsonable = False
 
@@ -1941,9 +1961,14 @@ class OrdinalEncoderTransformer(
     FITS: bool
         class attribute, indicates whether transform requires fit to be run first
 
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     """
 
     polars_compatible = False
+
+    lazyframe_compatible = False
 
     jsonable = False
 
@@ -2126,9 +2151,14 @@ class NominalToIntegerTransformer(BaseNominalTransformer, BaseMappingTransformMi
     FITS: bool
         class attribute, indicates whether transform requires fit to be run first
 
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     """
 
     polars_compatible = False
+
+    lazyframe_compatible = False
 
     jsonable = False