+++ tubular/numeric.py
@@ -71,6 +71,9 @@ class BaseNumericTransformer(BaseTransformer, CheckNumericMixin):
     jsonable: bool
         class attribute, indicates if transformer supports to/from_json methods
 
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     Example:
     --------
     >>> BaseNumericTransformer(
@@ -81,6 +84,8 @@ class BaseNumericTransformer(BaseTransformer, CheckNumericMixin):
     """
 
     polars_compatible = True
+
+    lazyframe_compatible = False
 
     jsonable = False
 
@@ -171,7 +176,7 @@ class BaseNumericTransformer(BaseTransformer, CheckNumericMixin):
         return_native = self._process_return_native(return_native_override)
         X = super().transform(X, return_native_override=False)
 
-        CheckNumericMixin.check_numeric_columns(self, X[self.columns])
+        CheckNumericMixin.check_numeric_columns(self, X.select(self.columns))
 
         return _return_narwhals_or_native_dataframe(X, return_native)
 
@@ -226,6 +231,9 @@ class OneDKmeansTransformer(BaseNumericTransformer, DropOriginalMixin):
     jsonable: bool
         class attribute, indicates if transformer supports to/from_json methods
 
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     Example:
     --------
     >>> OneDKmeansTransformer(
@@ -241,6 +249,8 @@ class OneDKmeansTransformer(BaseNumericTransformer, DropOriginalMixin):
     """
 
     polars_compatible = True
+
+    lazyframe_compatible = False
 
     jsonable = False
 
@@ -258,7 +268,7 @@ class OneDKmeansTransformer(BaseNumericTransformer, DropOriginalMixin):
         n_clusters: int = 8,
         drop_original: bool = False,
         kmeans_kwargs: Optional[dict[str, object]] = None,
-        **kwargs: dict[str, bool],
+        **kwargs: bool,
     ) -> None:
         if kmeans_kwargs is None:
             kmeans_kwargs = {}
@@ -450,6 +460,22 @@ class DifferenceTransformer(BaseNumericTransformer):
     columns : ListOfTwoStrs
         List of exactly two column names to operate on. The second column is subtracted from the first.
 
+    built_from_json: bool
+        indicates if transformer was reconstructed from json, which limits it's supported
+        functionality to .transform
+
+    polars_compatible : bool
+        class attribute, indicates whether transformer has been converted to polars/pandas agnostic narwhals framework
+
+    FITS: bool
+        class attribute, indicates whether transform requires fit to be run first
+
+    jsonable: bool
+        class attribute, indicates if transformer supports to/from_json methods
+
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     Example
     -------
     >>> transformer = DifferenceTransformer(columns=['a', 'b'])
@@ -458,8 +484,12 @@ class DifferenceTransformer(BaseNumericTransformer):
     """
 
     polars_compatible = True
+
     FITS = False
+
     jsonable = True
+
+    lazyframe_compatible = False
 
     @beartype
     def __init__(
@@ -552,6 +582,22 @@ class RatioTransformer(BaseNumericTransformer):
     return_dtype : str
         The dtype of the resulting column, either 'Float32' or 'Float64'.
 
+    built_from_json: bool
+        indicates if transformer was reconstructed from json, which limits it's supported
+        functionality to .transform
+
+    polars_compatible : bool
+        class attribute, indicates whether transformer has been converted to polars/pandas agnostic narwhals framework
+
+    FITS: bool
+        class attribute, indicates whether transform requires fit to be run first
+
+    jsonable: bool
+        class attribute, indicates if transformer supports to/from_json methods
+
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     Example
     -------
     >>> transformer = RatioTransformer(columns=['a', 'b'], return_dtype='Float32')
@@ -562,8 +608,12 @@ class RatioTransformer(BaseNumericTransformer):
     """
 
     polars_compatible = True
+
     FITS = False
+
     jsonable = True
+
+    lazyframe_compatible = False
 
     @block_from_json
     def to_json(self) -> dict[str, dict[str, Any]]:
@@ -729,9 +779,14 @@ class LogTransformer(BaseNumericTransformer, DropOriginalMixin):
     jsonable: bool
         class attribute, indicates if transformer supports to/from_json methods
 
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     """
 
     polars_compatible = False
+
+    lazyframe_compatible = False
 
     jsonable = False
 
@@ -745,7 +800,7 @@ class LogTransformer(BaseNumericTransformer, DropOriginalMixin):
         add_1: bool = False,
         drop_original: bool = True,
         suffix: str = "log",
-        **kwargs: dict[str, bool],
+        **kwargs: bool,
     ) -> None:
         super().__init__(columns=columns, **kwargs)
 
@@ -843,11 +898,14 @@ class CutTransformer(BaseNumericTransformer):
     jsonable: bool
         class attribute, indicates if transformer supports to/from_json methods
 
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     """
 
     polars_compatible = False
 
-    jsonable = False
+    lazyframe_compatible = False
 
     FITS = False
 
@@ -857,7 +915,7 @@ class CutTransformer(BaseNumericTransformer):
         column: str,
         new_column_name: str,
         cut_kwargs: Optional[GenericKwargs] = None,
-        **kwargs: dict[str, bool],
+        **kwargs: bool,
     ) -> None:
         if cut_kwargs is None:
             cut_kwargs = {}
@@ -882,10 +940,16 @@ class CutTransformer(BaseNumericTransformer):
         """
         X = super().transform(X)
 
-        X[self.new_column_name] = pd.cut(
-            X[self.columns[0]].to_numpy(),
-            **self.cut_kwargs,
-        )
+        # quick fix for empty frames, not spending much
+        # time on this as transformer is deprecated
+        if X.empty:
+            X[self.new_column_name] = pd.Series(dtype=float)
+
+        else:
+            X[self.new_column_name] = pd.cut(
+                X[self.columns[0]].to_numpy(),
+                **self.cut_kwargs,
+            )
 
         return X
 
@@ -957,9 +1021,14 @@ class TwoColumnOperatorTransformer(
     jsonable: bool
         class attribute, indicates if transformer supports to/from_json methods
 
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     """
 
     polars_compatible = False
+
+    lazyframe_compatible = False
 
     jsonable = False
 
@@ -1066,9 +1135,14 @@ class ScalingTransformer(BaseNumericTransformer):
     jsonable: bool
         class attribute, indicates if transformer supports to/from_json methods
 
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     """
 
     polars_compatible = False
+
+    lazyframe_compatible = False
 
     jsonable = False
 
@@ -1150,7 +1224,14 @@ class ScalingTransformer(BaseNumericTransformer):
         """
         X = super().transform(X)
 
-        X[self.columns] = self.scaler.transform(X[self.columns])
+        # quick fix for empty frames, not spending much
+        # time on this as transformer is deprecated
+        if X.empty:
+            for col in self.columns:
+                X[col] = pd.Series(dtype=float)
+
+        else:
+            X[self.columns] = self.scaler.transform(X[self.columns])
 
         return X
 
@@ -1222,9 +1303,14 @@ class InteractionTransformer(BaseNumericTransformer):
         jsonable: bool
             class attribute, indicates if transformer supports to/from_json methods
 
+        lazyframe_compatible: bool
+            class attribute, indicates whether transformer works with lazyframes
+
     """
 
     polars_compatible = False
+
+    lazyframe_compatible = False
 
     jsonable = False
 
@@ -1420,9 +1506,14 @@ class PCATransformer(BaseNumericTransformer):
     jsonable: bool
         class attribute, indicates if transformer supports to/from_json methods
 
+    lazyframe_compatible: bool
+        class attribute, indicates whether transformer works with lazyframes
+
     """
 
     polars_compatible = False
+
+    lazyframe_compatible = False
 
     jsonable = False
 
@@ -1543,6 +1634,14 @@ class PCATransformer(BaseNumericTransformer):
         """
         X = super().transform(X)
         X = CheckNumericMixin.check_numeric_columns(self, X)
-        X[self.feature_names_out] = self.pca.transform(X[self.columns])
+
+        # quick fix for empty frames, not spending much
+        # time on this as transformer is deprecated
+        if X.empty:
+            for col in self.feature_names_out:
+                X[col] = pd.Series(dtype=float)
+
+        else:
+            X[self.feature_names_out] = self.pca.transform(X[self.columns])
 
         return X